requirements.txt
chime==0.6.0
deta==1.1.0
fpdf==1.7.2
fpdf2==2.5.1
hydralit_components==1.0.9
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
Pillow==9.1.0
plotly==5.6.0
pydicom==2.2.2
pyecharts==1.9.1
pylinac==3.1.0a2
pyparsing==3.0.7
requests==2.27.1
SimpleITK==2.1.1
st_btn_select==0.1.2
streamlit==1.7.0
streamlit_authenticator==0.1.2
streamlit_autorefresh==0.0.1
streamlit_echarts==0.4.0
streamlit_lottie==0.0.3
click<=8.0.4
pytz==2021.3
